Updated en-us resources
