+++ SharePointFramework/ProjectWebParts/src/components/BaseWebPartComponent/types.ts
@@ -38,3 +38,25 @@ export interface IBaseWebPartComponentProps {
      */
     displayMode?: DisplayMode;
 }
+
+export interface IBaseWebPartComponentState<T> {
+    /**
+     * The component is loading
+     */
+    isLoading: boolean;
+
+    /**
+     * Data 
+     */
+    data?: T;
+
+    /**
+     * Error object
+     */
+    error?: any;
+
+    /**
+     * Is the component hidden
+     */
+    hidden?: boolean;
+}
