+++ SharePointFramework/ProjectWebParts/src/loc/en-us.js
@@ -17,8 +17,8 @@ define ([], function () {
     ConfirmPhaseDialogTitle: 'Change Phase',
     CurrentPhaseViewNameFieldLabel: 'Current phase display name',
     DocumentsListName: 'Documents',
-    EditPropertiesText: 'Edit Properties',
-    EditReportButtonText: 'Edit Status',
+    EditPropertiesText: 'Edit properties',
+    EditReportButtonText: 'Edit status',
     PublishReportButtonText: 'Publish',
     EditSiteInformationText: 'Edit site information',
     ErrorText: 'There was an error loading {0}.',
